Camera service: Speed up camera service startup on boot
In the common case, cameraserver starts before system server, which
contains the proxy camera service.  This means that
getCameraServiceProxy() will wait for the proxy to be up (or time out
after 5 seconds).  Since the proxy will call the camera service once it
is up and running, there's no reason to block.

Switch to checkService() instead of getService() to remove this timeout
and fail immediately.

Test: Manual: Build and boot phone; confirm that camera functions normally
Change-Id: I03596c18b13c68b5bcb15ee71de465cd14fba4de